extract view.ts from editor/modal.ts
Moves editorHtml template, MODE type, COLOR_PRESETS, escapeHtml, and
promptRestoreDraft into modal/view.ts. modal.ts imports from the new file.
